requirements.txt
aiosqlite==0.20.0
anyio==4.6.0
certifi==2024.8.30
flake8==7.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
mccabe==0.7.0
mypy==1.11.2
mypy-extensions==1.0.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
python-telegram-bot==21.6
sniffio==1.3.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
